Fixed failing tests - had to roll back some changes and also pulled out FreeFormObjectsTest
